Split ui/update.go into per-mode handler files (#17)
Move 14 handler functions out of the 1029-line update.go into four
focused files, matching the existing one-file-per-concern convention:

- update_normal.go:  handleNormalKey, switchToServer
- update_form.go:    handleFormKey, handleNewJobChoiceKey, handleTemplatePickerKey
- update_server.go:  handleAddServerKey, handlePasswordPromptKey, handleConfirmDeleteServerKey
- update_dialogs.go: handleConfirmKey, handleHelpKey, handleConfirmDeleteHistoryKey, handleRunOutputKey

update.go retains only the top-level Update() dispatcher and handleKey()
router. Pure file-move refactor with zero API or behavior changes.
